Add files via upload
This is a simple slam "simulator" for Thunder Road: Vendetta. When slamming into an opponent (or even yourself), you have to roll a few dice to see what player will move, and what tile they will move to. I thought it would be interesting to code this as my first small project. Also had to force myself to stop because I keep trying to add or change small details that shouldn't matter at this point in time.